@@ -21,26 +21,28 @@
     "build:prod": "ng build --configuration production",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "dev:ssr": "ng run jewelry-store:serve-ssr",
+    "dev:ssr": "ng run jewelry-store:serve-ssr:development",
     "serve:ssr": "node dist/jewelry-store/server/main.js",
     "build:dev:ssr": "ng build --configuration=development && ng run jewelry-store:server",
     "build:prod:ssr": "ng build --configuration=production && ng run jewelry-store:server",
-    "prerender": "ng run jewelry-store:prerender"
+    "prerender": "ng run jewelry-store:prerender",
+    "build:ssr": "ng build && ng run jewelry-store:server"
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^17.0.6",
-    "@angular/animations": "^17.0.6",
-    "@angular/common": "^17.0.6",
-    "@angular/compiler": "^17.0.6",
-    "@angular/core": "^17.0.6",
-    "@angular/forms": "^17.0.6",
-    "@angular/platform-browser": "^17.0.6",
-    "@angular/platform-browser-dynamic": "^17.0.6",
-    "@angular/platform-server": "^17.0.6",
-    "@angular/router": "^17.0.6",
-    "@angular/ssr": "^17.0.6",
-    "@angular/youtube-player": "^17.0.3",
+    "@angular-devkit/build-angular": "^18.2.5",
+    "@angular-slider/ngx-slider": "^18.0.0",
+    "@angular/animations": "^18.2.5",
+    "@angular/common": "^18.2.5",
+    "@angular/compiler": "^18.2.5",
+    "@angular/core": "^18.2.5",
+    "@angular/forms": "^18.2.5",
+    "@angular/platform-browser": "^18.2.5",
+    "@angular/platform-browser-dynamic": "^18.2.5",
+    "@angular/platform-server": "^18.2.5",
+    "@angular/router": "^18.2.5",
+    "@angular/ssr": "^18.2.5",
+    "@angular/youtube-player": "^18.2.4",
     "@capacitor/android": "^6.0.0",
     "@capacitor/angular": "^2.0.3",
     "@capacitor/app": "^6.0.0",
@@ -52,39 +54,43 @@
     "@capacitor/splash-screen": "^6.0.0",
     "@capacitor/status-bar": "^6.0.0",
     "@ionic/angular": "^7.6.0",
-    "@ngrx/component-store": "^17.0.1",
+    "@ngrx/component-store": "^18.0.2",
     "@ngx-translate/core": "^14.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
     "angular-star-rating": "^6.0.0",
     "bootstrap": "^5.2.3",
+    "cookie-parser": "^1.4.6",
     "css-star-rating": "^1.3.1",
     "emulator": "0.1.0",
-    "express": "^4.15.2",
+    "express": "^4.18.2",
     "localstorage-polyfill": "^1.0.1",
     "lodash": "^4.17.21",
-    "ngx-bootstrap": "12.0.0",
-    "ngx-slider-v2": "^17.0.0",
-    "ngx-toastr": "^18.0.0",
+    "ngx-bootstrap": "18.0.2",
+    "ngx-cookie-service": "^18.0.0",
+    "ngx-cookie-service-ssr": "^18.0.0",
+    "ngx-toastr": "^19.0.0",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular/cli": "~17.0.6",
-    "@angular/compiler-cli": "^17.0.6",
+    "@angular/cli": "~18.2.5",
+    "@angular/compiler-cli": "^18.2.5",
     "@capacitor/assets": "^3.0.4",
     "@capacitor/cli": "^6.0.0",
     "@ionic/angular-toolkit": "^10.0.0",
-    "@types/express": "^4.17.0",
+    "@types/cookie-parser": "^1.4.7",
+    "@types/express": "^4.17.17",
     "@types/jasmine": "~4.3.0",
     "@types/lodash": "^4.14.202",
-    "@types/node": "^14.15.0",
+    "@types/node": "^22.5.5",
+    "browser-sync": "^3.0.0",
     "jasmine-core": "~4.5.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~5.2.2"
+    "typescript": "~5.5.4"
   }
 }
