@@ -1,15 +1,23 @@
-import { Injectable } from '@angular/core';
+import {Inject, Injectable, PLATFORM_ID} from '@angular/core';
 import { ActivatedRouteSnapshot, Router, RouterStateSnapshot, UrlSegment, UrlTree } from '@angular/router';
-import { map, Observable, of } from 'rxjs';
-import { defaultLanguage } from "../../constants/languages";
+import {map, Observable, of, take} from 'rxjs';
+import { selectedLanguage } from "../../constants/languages";
 import { TranslateService } from "@ngx-translate/core";
+import {isPlatformBrowser} from "@angular/common";
+import {LocalStorageService} from "@services/local-storage.service";
 
 @Injectable({
   providedIn: 'root'
 })
 export class LanguageGuard  {
 
-  constructor(private router: Router, private translateService: TranslateService) {
+  constructor(
+    private router: Router,
+    private translateService: TranslateService,
+    @Inject(PLATFORM_ID)
+    private platformId: Object,
+    private localStorageService: LocalStorageService
+  ) {
   }
 
   canActivate(
@@ -18,18 +26,29 @@ export class LanguageGuard  {
   ): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
     if(!this.translateService.langs.some(lang => lang === route.params["language"])) {
       const url = [...route.url];
-      url[0] = new UrlSegment(defaultLanguage, {});
+      url[0] = new UrlSegment(selectedLanguage, {});
       this.router.navigate(url, {
         queryParams: route.queryParams
       });
 
-      this.translateService.use(defaultLanguage).subscribe();
-      localStorage.setItem("lang", defaultLanguage);
+      this.translateService.use(selectedLanguage)
+        .pipe(
+          take(1)
+        ).subscribe();
+      if (isPlatformBrowser(this.platformId)) {
+        this.localStorageService.set("lang", selectedLanguage);
+      }
       return false;
     }
 
-    this.translateService.use(route.params["language"]).subscribe();
-    localStorage.setItem("lang", route.params["language"]);
+    if (isPlatformBrowser(this.platformId)) {
+      this.translateService.use(route.params["language"])
+        .pipe(
+          take(1)
+        )
+        .subscribe();
+      this.localStorageService.set("lang", route.params["language"]);
+    }
     return true;
   }
 
