@@ -1,10 +1,8 @@
 import {Inject, inject, Injectable} from '@angular/core';
-import {ComponentStore, OnStoreInit, tapResponse} from "@ngrx/component-store";
+import {ComponentStore, OnStoreInit} from "@ngrx/component-store";
 import {catchError, EMPTY, filter, map, Observable, switchMap, take, takeUntil, tap} from "rxjs";
 import {ToastrService} from "ngx-toastr";
 import {LocalStorageService} from "@services/local-storage.service";
-import {localStorageFactory} from "@factories/local-storage.factory";
-import {LOCAL_STORAGE_TOKEN} from "../injection-tokens/local-storage.token";
 import {Product} from "@interfaces/product";
 import {FavoritesService} from "@services/favorites.service";
 
@@ -112,15 +110,14 @@ export class FavoritesStore extends ComponentStore<FavoritesStoreState> implemen
         take(1)
       )),
       switchMap((ids) => this.favoritesService.getFavoriteProducts([...ids])),
-      tapResponse(
-        (response) => {
-          console.log(response);
+      tap({
+        next: (response) => {
           this.setFavoriteProductsSuccess(response);
         },
-        (error) => {
+        error: (error) => {
           this.setFavoriteProductsFailure(error);
         }
-      )
+      })
     )
   })
 
